@AngersZhuuuu 2020-11-30
> It's a behavior change which could be OK if it's a bug fix. Do SQL standards have anything to say about what the result should be?

Didn't  find the place to describe this exactly in SQL standards. It's a function from hive and I prepare to make a pr in hive too.
From the benchmark, seems  have little influence in performance.