@cloud-fan 2020-12-01
Fixing overflow should be good. It's not very intuitive to read the results of this function. Is there any database we can use as a reference?