@AngersZhuuuu 2020-12-03
> Fixing overflow should be good. It's not very intuitive to read the results of this function. Is there any database we can use as a reference?

Mysql: https://dev.mysql.com/doc/refman/8.0/en/mathematical-functions.html#function_conv
But still only support 64-bit presicion.

Hive: https://cwiki.apache.org/confluence/display/Hive/LanguageManual+UDF
Latest version support BigInt now
![image](https://user-images.githubusercontent.com/46485123/100971384-c0f83280-3571-11eb-9747-b60f2c489010.png)