@cloud-fan 2020-12-14
the scala 2.13 build is broken:
```
[error] /home/runner/work/spark/spark/sql/catalyst/src/main/scala/org/apache/spark/sql/catalyst/util/NumberConverter.scala:36:46: Byte does not take parameters
[error]       value(i) = (tmpV - q * radix).byteValue()
[error]                                      
```