@AngersZhuuuu 2020-12-14
> value(i) = (tmpV - q * radix).byteValue()

Thanks for mention this point