@cloud-fan 2020-12-14
Since the overflow behavior is expected, this is more like an improvement instead of a bug fix. Thus, I'm merging it to master only, thanks!