@cloud-fan 2021-03-15
I found a behavior change here.

Before this commit
```
scala> sql("select conv(-100, 10, 10)").show(20, false)
+----------------------------------+
|conv(CAST(-100 AS STRING), 10, 10)|
+----------------------------------+
|18446744073709551516              |
+----------------------------------+
```

After this commit
```
scala> sql("select conv(-100, 10, 10)").show(20, false)
+------------------+
|conv(-100, 10, 10)|
+------------------+
|-100              |
+------------------+
```

If the `toBase` is positive, the result is unsigned. It's a bit hard to reason about how to put a negative value into an unsigned value, but MySQL returns the same result as before this commit and the overflow behavior seems to be desirable. @AngersZhuuuu can you take a look?