@AngersZhuuuu 2021-03-16
Yea, check the detail of mysql https://www.w3resource.com/mysql/mathematical-functions/mysql-conv-function.php
I means if to base is positive, we treat it as unsigned. then when overflow, it return `-1`(unsigned long value as `18446744073709551615` ),  for current mysql behavior, 

1. User always not clear about the overflow, so if we return overflow  value, user may not known it.
2. Long can't satisfy current user's requirement, if we use a big decimal or bigint, we can't use this fuction.
3. Keep same with other engine is good.
4. For both keep same and satisfy user's requirement,  should we revert this change and add a new function such as `conv_without_overflow`?