@cloud-fan 2021-03-16
I've reverted this commit.

Regarding `conv_without_overflow`, are we going to disallow negative `toBase`? Does any other database have it?