@AngersZhuuuu 2021-03-16
> I've reverted this commit.
> 
> Regarding `conv_without_overflow`, are we going to disallow negative `toBase`? Does any other database have it?

Hmmm why do we need to disallow negative `toBase`. Doesn't find other database have it.