@cloud-fan 2021-03-16
If we never overflow, why do we need to return unsigned results?