@AngersZhuuuu 2021-03-16
> If we never overflow, why do we need to return unsigned results?

Hmmm you mean negative number? `conv(number, fromBase, toBase)`. This pr won't change behavior when negative `toBase`