@cloud-fan 2021-03-16
Do we still need this signed vs unsigned difference in the `conv` function if it never overflows?